64-bit sim fixes
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@9316 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/doom/p_saveg.c
@@ -44,7 +44,7 @@ byte *save_p;
 
 // Pads save_p to a 4-byte boundary
 //  so that the load/save works on SGI&Gecko.
-#define PADSAVEP()    do { save_p += (4 - ((int) save_p & 3)) & 3; } while (0)
+#define PADSAVEP()    do { save_p += (4 - ((unsigned long) save_p & 3)) & 3; } while (0)
 //
 // P_ArchivePlayers
 //
